requirements.txt
django>=3.1.1,<3.2.0
djangorestframework>=3.11.1,<3.12.0
requests>=2.24.0,<2.25.0
redis>=3.5.3,<3.6.0
locust==1.2.3